Rescheduled  validation schemas
1.) Rescheduled validation schemas in validation/validationJoiSchemas.js;
2.) Сode refactoring
3.) Renamed route file
4.)Tested on postman
